requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
durationpy==0.9
google-auth==2.39.0
idna==3.10
kubernetes==32.0.1
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
urllib3==2.4.0
websocket-client==1.8.0
